update tests and rename components
